requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
phonenumbers>=8.13.0
playwright>=1.44.0
ddgs>=0.1.0
tqdm>=4.0.0
openai>=1.30.0
python-dotenv>=1.0.0
streamlit>=1.35.0
